Refactor application structure
