+++ src/PieShop.Web/Core/IFeedbackRepository.cs
@@ -1,0 +1,9 @@
+﻿using PieShop.Web.Models;
+
+namespace PieShop.Web.Core
+{
+    public interface IFeedbackRepository
+    {
+        void AddFeedback(Feedback feedback);
+    }
+}
