--- src/PieShop.Web/Models/IFeedbackRepository.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace PieShop.Web.Models
-{
-    public interface IFeedbackRepository
-    {
-        void AddFeedback(Feedback feedback);
-    }
-}
