+++ src/PieShop.Web/ViewModels/HomeViewModel.cs
@@ -1,8 +1,5 @@
 ﻿using PieShop.Web.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace PieShop.Web.ViewModels
 {
