+++ src/PieShop.Web/appsettings.json
@@ -1,5 +1,6 @@
 ﻿{
   "ConnectionStrings": {
+    //Enter your connection string here
     "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=PieShop;Trusted_Connection=True;MultipleActiveResultSets=true"
   }
 }
